Remove libspotify dependency
The web API can now delete tracks from playlists. This means that
libspotify is only providing one feature that cannot be replaced by the
web API: deleting local tracks (that don't have a Spotify ID) from
playlists.

The usefulness of that feature is outweighed by the inconvenience of
libspotify - it requires a premium account, you have to manage a
separate set of API credentials, and it takes longer to install.

Remove libspotify, replacing the track-deletion features with the
spotipy client.